m3t1 is started
